requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.3
Jinja2==3.1.4
Authlib==1.3.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-dotenv==1.0.1
itsdangerous==2.1.2
httpx==0.25.0
PyJWT==2.8.0
